requirements.txt
Flask>=2.3.0
Pillow>=10.0.0
python-docx>=0.8.11
pandas>=2.0.0
requests>=2.31.0
lxml>=4.9.0
python-dateutil>=2.8.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
alembic==1.13.1

# Caching
redis==5.0.1
flask-caching==2.1.0

# Authentication
flask-jwt-extended==4.6.0
bcrypt==4.1.2
werkzeug==3.1.1

# Validation
marshmallow==3.20.2
flask-marshmallow==0.15.0

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
